Refactor CLI frontend parsing into dedicated modules
