+++ crates/cli/src/frontend/mod.rs
@@ -94,19 +94,18 @@ use std::str::FromStr;
 use crate::platform::{gid_t, uid_t};
 mod command_builder;
 
-use command_builder::clap_command;
+use args::{BandwidthArgument, ParsedArgs, parse_args};
 use rsync_compress::zlib::CompressionLevel;
 #[cfg(test)]
 use rsync_core::client::DirMergeEnforcedKind;
 use rsync_core::{
     bandwidth::BandwidthParseError,
-    branding::{self, Brand},
+    branding::Brand,
     client::{
-        AddressMode, BandwidthLimit, BindAddress, ClientConfig, ClientOutcome,
-        ClientProgressObserver, CompressionSetting, DeleteMode, DirMergeOptions, FilterRuleKind,
-        FilterRuleSpec, HumanReadableMode, ModuleListOptions, ModuleListRequest,
-        RemoteFallbackArgs, RemoteFallbackContext, StrongChecksumChoice, TransferTimeout,
-        parse_skip_compress_list, run_client_or_fallback,
+        BandwidthLimit, BindAddress, ClientConfig, ClientOutcome, ClientProgressObserver,
+        CompressionSetting, DeleteMode, DirMergeOptions, FilterRuleKind, FilterRuleSpec,
+        HumanReadableMode, ModuleListOptions, ModuleListRequest, RemoteFallbackArgs,
+        RemoteFallbackContext, TransferTimeout, parse_skip_compress_list, run_client_or_fallback,
         run_module_list_with_password_and_options, skip_compress_from_env,
     },
     message::{Message, Role},
@@ -116,11 +115,13 @@ use rsync_core::{
 use rsync_logging::MessageSink;
 use rsync_meta::ChmodModifiers;
 use rsync_protocol::{ParseProtocolVersionErrorKind, ProtocolVersion};
+mod args;
 mod defaults;
 mod filter_rules;
 mod help;
 mod out_format;
 pub(crate) mod password;
+mod program;
 mod progress;
 mod server;
 
@@ -131,12 +132,12 @@ pub(crate) use defaults::LIST_TIMESTAMP_FORMAT;
 use defaults::{ITEMIZE_CHANGES_FORMAT, SUPPORTED_OPTIONS_LIST};
 use filter_rules::{
     FilterDirective, append_cvs_exclude_rules, append_filter_rules_from_files,
-    apply_merge_directive, collect_filter_arguments, locate_filter_arguments,
-    merge_directive_options, os_string_to_pattern, parse_filter_directive,
+    apply_merge_directive, merge_directive_options, os_string_to_pattern, parse_filter_directive,
 };
 use help::help_text;
 pub(crate) use out_format::{OutFormat, OutFormatContext, emit_out_format, parse_out_format};
 use password::{load_optional_password, load_password_file};
+use program::{ProgramName, detect_program_name};
 pub(crate) use progress::*;
 
 #[cfg(test)]
@@ -182,748 +183,6 @@ pub(crate) fn process_merge_directive(
 
 /// Maximum exit code representable by a Unix process.
 const MAX_EXIT_CODE: i32 = u8::MAX as i32;
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ProgramName {
-    Rsync,
-    OcRsync,
-}
-
-impl ProgramName {
-    #[inline]
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Rsync => Brand::Upstream.client_program_name(),
-            Self::OcRsync => Brand::Oc.client_program_name(),
-        }
-    }
-
-    #[inline]
-    const fn brand(self) -> Brand {
-        match self {
-            Self::Rsync => Brand::Upstream,
-            Self::OcRsync => Brand::Oc,
-        }
-    }
-}
-
-fn detect_program_name(program: Option<&OsStr>) -> ProgramName {
-    match branding::detect_brand(program) {
-        Brand::Oc => ProgramName::OcRsync,
-        Brand::Upstream => ProgramName::Rsync,
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-enum BandwidthArgument {
-    Limit(OsString),
-    Disabled,
-}
-
-struct ParsedArgs {
-    program_name: ProgramName,
-    show_help: bool,
-    show_version: bool,
-    human_readable: Option<HumanReadableMode>,
-    dry_run: bool,
-    list_only: bool,
-    remote_shell: Option<OsString>,
-    connect_program: Option<OsString>,
-    remote_options: Vec<OsString>,
-    rsync_path: Option<OsString>,
-    protect_args: Option<bool>,
-    address_mode: AddressMode,
-    bind_address: Option<OsString>,
-    archive: bool,
-    delete_mode: DeleteMode,
-    delete_excluded: bool,
-    backup: bool,
-    backup_dir: Option<OsString>,
-    backup_suffix: Option<OsString>,
-    checksum: bool,
-    checksum_choice: Option<StrongChecksumChoice>,
-    checksum_choice_arg: Option<OsString>,
-    checksum_seed: Option<u32>,
-    size_only: bool,
-    ignore_existing: bool,
-    ignore_missing_args: bool,
-    update: bool,
-    remainder: Vec<OsString>,
-    bwlimit: Option<BandwidthArgument>,
-    max_delete: Option<OsString>,
-    min_size: Option<OsString>,
-    max_size: Option<OsString>,
-    modify_window: Option<OsString>,
-    compress: bool,
-    no_compress: bool,
-    compress_level: Option<OsString>,
-    skip_compress: Option<OsString>,
-    owner: Option<bool>,
-    group: Option<bool>,
-    chown: Option<OsString>,
-    chmod: Vec<OsString>,
-    perms: Option<bool>,
-    super_mode: Option<bool>,
-    times: Option<bool>,
-    omit_dir_times: Option<bool>,
-    omit_link_times: Option<bool>,
-    acls: Option<bool>,
-    numeric_ids: Option<bool>,
-    hard_links: Option<bool>,
-    sparse: Option<bool>,
-    copy_links: Option<bool>,
-    copy_dirlinks: bool,
-    copy_unsafe_links: Option<bool>,
-    keep_dirlinks: Option<bool>,
-    safe_links: bool,
-    devices: Option<bool>,
-    specials: Option<bool>,
-    relative: Option<bool>,
-    one_file_system: Option<bool>,
-    implied_dirs: Option<bool>,
-    mkpath: bool,
-    prune_empty_dirs: Option<bool>,
-    verbosity: u8,
-    progress: ProgressSetting,
-    name_level: NameOutputLevel,
-    name_overridden: bool,
-    stats: bool,
-    partial: bool,
-    preallocate: bool,
-    delay_updates: bool,
-    partial_dir: Option<PathBuf>,
-    temp_dir: Option<PathBuf>,
-    link_dests: Vec<PathBuf>,
-    remove_source_files: bool,
-    inplace: Option<bool>,
-    append: Option<bool>,
-    append_verify: bool,
-    msgs_to_stderr: bool,
-    itemize_changes: bool,
-    whole_file: Option<bool>,
-    excludes: Vec<OsString>,
-    includes: Vec<OsString>,
-    exclude_from: Vec<OsString>,
-    include_from: Vec<OsString>,
-    filters: Vec<OsString>,
-    compare_destinations: Vec<OsString>,
-    copy_destinations: Vec<OsString>,
-    link_destinations: Vec<OsString>,
-    cvs_exclude: bool,
-    rsync_filter_shortcuts: u8,
-    files_from: Vec<OsString>,
-    from0: bool,
-    info: Vec<OsString>,
-    debug: Vec<OsString>,
-    xattrs: Option<bool>,
-    no_motd: bool,
-    password_file: Option<OsString>,
-    protocol: Option<OsString>,
-    timeout: Option<OsString>,
-    contimeout: Option<OsString>,
-    out_format: Option<OsString>,
-    daemon_port: Option<u16>,
-}
-
-fn env_protect_args_default() -> Option<bool> {
-    let value = env::var_os("RSYNC_PROTECT_ARGS")?;
-    if value.is_empty() {
-        return Some(true);
-    }
-
-    let normalized = value.to_string_lossy();
-    let trimmed = normalized.trim();
-
-    if trimmed.is_empty() {
-        Some(true)
-    } else if trimmed.eq_ignore_ascii_case("0")
-        || trimmed.eq_ignore_ascii_case("no")
-        || trimmed.eq_ignore_ascii_case("false")
-        || trimmed.eq_ignore_ascii_case("off")
-    {
-        Some(false)
-    } else {
-        Some(true)
-    }
-}
-
-/// Builds the `clap` command used for parsing.
-/// Parses command-line arguments into a [`ParsedArgs`] structure.
-fn parse_args<I, S>(arguments: I) -> Result<ParsedArgs, clap::Error>
-where
-    I: IntoIterator<Item = S>,
-    S: Into<OsString>,
-{
-    let mut args: Vec<OsString> = arguments.into_iter().map(Into::into).collect();
-
-    let program_name = detect_program_name(args.first().map(OsString::as_os_str));
-
-    if args.is_empty() {
-        args.push(OsString::from(program_name.as_str()));
-    }
-
-    let raw_args = args.clone();
-    let (filter_indices, rsync_filter_indices) = locate_filter_arguments(&raw_args);
-    let mut matches = clap_command(program_name.as_str()).try_get_matches_from(args)?;
-
-    let show_help = matches.get_flag("help");
-    let show_version = matches.get_flag("version");
-    let mut human_readable = matches
-        .remove_one::<OsString>("human-readable")
-        .map(|value| parse_human_readable_level(value.as_os_str()))
-        .transpose()?;
-    if matches.get_flag("no-human-readable") {
-        human_readable = Some(HumanReadableMode::Disabled);
-    }
-    let mut dry_run = matches.get_flag("dry-run");
-    let list_only = matches.get_flag("list-only");
-    let mkpath = matches.get_flag("mkpath");
-    let prune_empty_dirs = if matches.get_flag("no-prune-empty-dirs") {
-        Some(false)
-    } else if matches.get_flag("prune-empty-dirs") {
-        Some(true)
-    } else {
-        None
-    };
-    let omit_link_times = if matches.get_flag("no-omit-link-times") {
-        Some(false)
-    } else if matches.get_flag("omit-link-times") {
-        Some(true)
-    } else {
-        None
-    };
-    if list_only {
-        dry_run = true;
-    }
-    let remote_shell = matches
-        .remove_one::<OsString>("rsh")
-        .filter(|value| !value.is_empty())
-        .or_else(|| env::var_os("RSYNC_RSH").filter(|value| !value.is_empty()));
-    let rsync_path = matches
-        .remove_one::<OsString>("rsync-path")
-        .filter(|value| !value.is_empty());
-    let connect_program = matches
-        .remove_one::<OsString>("connect-program")
-        .filter(|value| !value.is_empty());
-    let daemon_port = matches.remove_one::<u16>("port");
-    let remote_options = matches
-        .remove_many::<OsString>("remote-option")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let protect_args = if matches.get_flag("no-protect-args") {
-        Some(false)
-    } else if matches.get_flag("protect-args") {
-        Some(true)
-    } else {
-        env_protect_args_default()
-    };
-    let address_mode = if matches.get_flag("ipv4") {
-        AddressMode::Ipv4
-    } else if matches.get_flag("ipv6") {
-        AddressMode::Ipv6
-    } else {
-        AddressMode::Default
-    };
-    let bind_address_raw = matches.remove_one::<OsString>("address");
-    let archive = matches.get_flag("archive");
-    let delete_flag = matches.get_flag("delete");
-    let delete_before_flag = matches.get_flag("delete-before");
-    let delete_during_flag = matches.get_flag("delete-during");
-    let delete_delay_flag = matches.get_flag("delete-delay");
-    let delete_after_flag = matches.get_flag("delete-after");
-    let ignore_missing_args = matches.get_flag("ignore-missing-args");
-    let delete_excluded = matches.get_flag("delete-excluded");
-    let max_delete = matches.remove_one::<OsString>("max-delete");
-    let min_size = matches.remove_one::<OsString>("min-size");
-    let max_size = matches.remove_one::<OsString>("max-size");
-    let modify_window = matches.remove_one::<OsString>("modify-window");
-
-    let delete_mode_conflicts = [
-        delete_before_flag,
-        delete_during_flag,
-        delete_delay_flag,
-        delete_after_flag,
-    ]
-    .into_iter()
-    .filter(|flag| *flag)
-    .count();
-
-    if delete_mode_conflicts > 1 {
-        return Err(clap::Error::raw(
-            clap::error::ErrorKind::ArgumentConflict,
-            "--delete-before, --delete-during, --delete-delay, and --delete-after are mutually exclusive",
-        ));
-    }
-
-    let mut delete_mode = if delete_before_flag {
-        DeleteMode::Before
-    } else if delete_delay_flag {
-        DeleteMode::Delay
-    } else if delete_after_flag {
-        DeleteMode::After
-    } else if delete_during_flag || delete_flag {
-        DeleteMode::During
-    } else {
-        DeleteMode::Disabled
-    };
-
-    if delete_excluded && !delete_mode.is_enabled() {
-        delete_mode = DeleteMode::During;
-    }
-    if max_delete.is_some() && !delete_mode.is_enabled() {
-        delete_mode = DeleteMode::During;
-    }
-    let mut backup = matches.get_flag("backup");
-    let backup_dir = matches.remove_one::<OsString>("backup-dir");
-    let backup_suffix = matches.remove_one::<OsString>("suffix");
-    if backup_dir.is_some() || backup_suffix.is_some() {
-        backup = true;
-    }
-    let compress_flag = matches.get_flag("compress");
-    let no_compress = matches.get_flag("no-compress");
-    let mut compress = if no_compress { false } else { compress_flag };
-    let compress_level_opt = matches.get_one::<OsString>("compress-level").cloned();
-    if let Some(ref value) = compress_level_opt {
-        if let Ok(setting) = parse_compress_level_argument(value.as_os_str()) {
-            compress = !setting.is_disabled();
-        }
-    }
-    let owner = if matches.get_flag("owner") {
-        Some(true)
-    } else if matches.get_flag("no-owner") {
-        Some(false)
-    } else {
-        None
-    };
-    let group = if matches.get_flag("group") {
-        Some(true)
-    } else if matches.get_flag("no-group") {
-        Some(false)
-    } else {
-        None
-    };
-    let chown = matches.remove_one::<OsString>("chown");
-    let chmod = matches
-        .remove_many::<OsString>("chmod")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let perms = if matches.get_flag("perms") {
-        Some(true)
-    } else if matches.get_flag("no-perms") {
-        Some(false)
-    } else {
-        None
-    };
-    let super_mode = if matches.get_flag("super") {
-        Some(true)
-    } else if matches.get_flag("no-super") {
-        Some(false)
-    } else {
-        None
-    };
-    let times = if matches.get_flag("times") {
-        Some(true)
-    } else if matches.get_flag("no-times") {
-        Some(false)
-    } else {
-        None
-    };
-    let omit_dir_times = if matches.get_flag("omit-dir-times") {
-        Some(true)
-    } else if matches.get_flag("no-omit-dir-times") {
-        Some(false)
-    } else {
-        None
-    };
-    let acls = if matches.get_flag("acls") {
-        Some(true)
-    } else if matches.get_flag("no-acls") {
-        Some(false)
-    } else {
-        None
-    };
-    let xattrs = if matches.get_flag("xattrs") {
-        Some(true)
-    } else if matches.get_flag("no-xattrs") {
-        Some(false)
-    } else {
-        None
-    };
-    let numeric_ids = if matches.get_flag("numeric-ids") {
-        Some(true)
-    } else if matches.get_flag("no-numeric-ids") {
-        Some(false)
-    } else {
-        None
-    };
-    let hard_links = if matches.get_flag("no-hard-links") {
-        Some(false)
-    } else if matches.get_flag("hard-links") {
-        Some(true)
-    } else {
-        None
-    };
-    let sparse = if matches.get_flag("sparse") {
-        Some(true)
-    } else if matches.get_flag("no-sparse") {
-        Some(false)
-    } else {
-        None
-    };
-    let copy_links = if matches.get_flag("no-copy-links") {
-        Some(false)
-    } else if matches.get_flag("copy-links") {
-        Some(true)
-    } else {
-        None
-    };
-    let copy_unsafe_links_option = if matches.get_flag("no-copy-unsafe-links") {
-        Some(false)
-    } else if matches.get_flag("copy-unsafe-links") {
-        Some(true)
-    } else {
-        None
-    };
-    let copy_dirlinks = matches.get_flag("copy-dirlinks");
-    let keep_dirlinks = if matches.get_flag("no-keep-dirlinks") {
-        Some(false)
-    } else if matches.get_flag("keep-dirlinks") {
-        Some(true)
-    } else {
-        None
-    };
-    let mut safe_links = matches.get_flag("safe-links");
-    if copy_unsafe_links_option == Some(true) {
-        safe_links = true;
-    }
-    let archive_devices = matches.get_flag("archive-devices");
-    let devices = if matches.get_flag("no-devices") {
-        Some(false)
-    } else if matches.get_flag("devices") || archive_devices {
-        Some(true)
-    } else {
-        None
-    };
-    let specials = if matches.get_flag("no-specials") {
-        Some(false)
-    } else if matches.get_flag("specials") || archive_devices {
-        Some(true)
-    } else {
-        None
-    };
-    let relative = if matches.get_flag("relative") {
-        Some(true)
-    } else if matches.get_flag("no-relative") {
-        Some(false)
-    } else {
-        None
-    };
-    let one_file_system = if matches.get_flag("no-one-file-system") {
-        Some(false)
-    } else if matches.get_flag("one-file-system") {
-        Some(true)
-    } else {
-        None
-    };
-    let implied_dirs = if matches.get_flag("implied-dirs") {
-        Some(true)
-    } else if matches.get_flag("no-implied-dirs") {
-        Some(false)
-    } else {
-        None
-    };
-    let verbosity = matches.get_count("verbose") as u8;
-    let name_level = match verbosity {
-        0 => NameOutputLevel::Disabled,
-        1 => NameOutputLevel::UpdatedOnly,
-        _ => NameOutputLevel::UpdatedAndUnchanged,
-    };
-    let name_overridden = false;
-    let progress_flag = matches.get_flag("progress");
-    let no_progress_flag = matches.get_flag("no-progress");
-    let mut progress_setting = if progress_flag {
-        ProgressSetting::PerFile
-    } else {
-        ProgressSetting::Unspecified
-    };
-    let stats = matches.get_flag("stats");
-    let mut partial = matches.get_flag("partial");
-    let preallocate = matches.get_flag("preallocate");
-    let mut delay_updates = matches.get_flag("delay-updates");
-    let mut partial_dir = matches
-        .remove_one::<OsString>("partial-dir")
-        .map(PathBuf::from);
-    let temp_dir = matches
-        .remove_one::<OsString>("temp-dir")
-        .map(PathBuf::from);
-    if partial_dir.is_some() {
-        partial = true;
-    }
-    if matches.get_flag("no-delay-updates") {
-        delay_updates = false;
-    }
-    if no_progress_flag {
-        progress_setting = ProgressSetting::Disabled;
-    }
-    if matches.get_flag("no-partial") {
-        partial = false;
-        partial_dir = None;
-        delay_updates = false;
-    }
-    if matches.get_count("partial-progress") > 0 {
-        partial = true;
-        if !no_progress_flag {
-            progress_setting = ProgressSetting::PerFile;
-        }
-    }
-    let link_dests: Vec<PathBuf> = matches
-        .remove_many::<OsString>("link-dest")
-        .map(|values| {
-            values
-                .filter(|value| !value.is_empty())
-                .map(PathBuf::from)
-                .collect()
-        })
-        .unwrap_or_default();
-    let link_destinations: Vec<OsString> = link_dests
-        .iter()
-        .map(|path| path.clone().into_os_string())
-        .collect();
-    let remove_source_files =
-        matches.get_flag("remove-source-files") || matches.get_flag("remove-sent-files");
-    let append_verify_flag = matches.get_flag("append-verify");
-    let append_flag = matches.get_flag("append");
-    let no_append_flag = matches.get_flag("no-append");
-    let append = if append_verify_flag || append_flag {
-        Some(true)
-    } else if no_append_flag {
-        Some(false)
-    } else {
-        None
-    };
-    let inplace = if matches.get_flag("no-inplace") {
-        Some(false)
-    } else if matches.get_flag("inplace") {
-        Some(true)
-    } else {
-        None
-    };
-    let msgs_to_stderr = matches.get_flag("msgs2stderr");
-    let whole_file = if matches.get_flag("whole-file") {
-        Some(true)
-    } else if matches.get_flag("no-whole-file") {
-        Some(false)
-    } else {
-        None
-    };
-    let remainder = matches
-        .remove_many::<OsString>("args")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let checksum = matches.get_flag("checksum");
-    let size_only = matches.get_flag("size-only");
-    let (checksum_choice, checksum_choice_arg) =
-        match matches.remove_one::<OsString>("checksum-choice") {
-            Some(value) => {
-                let text = value.to_string_lossy().into_owned();
-                match StrongChecksumChoice::parse(&text) {
-                    Ok(choice) => {
-                        let normalized = OsString::from(choice.to_argument());
-                        (Some(choice), Some(normalized))
-                    }
-                    Err(message) => {
-                        return Err(clap::Error::raw(
-                            clap::error::ErrorKind::ValueValidation,
-                            message.text().to_string(),
-                        ));
-                    }
-                }
-            }
-            None => (None, None),
-        };
-
-    let checksum_seed = match matches.remove_one::<OsString>("checksum-seed") {
-        Some(value) => match parse_checksum_seed_argument(value.as_os_str()) {
-            Ok(seed) => Some(seed),
-            Err(message) => {
-                return Err(clap::Error::raw(
-                    clap::error::ErrorKind::ValueValidation,
-                    message.text().to_string(),
-                ));
-            }
-        },
-        None => None,
-    };
-
-    let compress_level = matches.remove_one::<OsString>("compress-level");
-    let skip_compress = matches.remove_one::<OsString>("skip-compress");
-    let no_bwlimit = matches.get_flag("no-bwlimit");
-    let bwlimit = if no_bwlimit {
-        Some(BandwidthArgument::Disabled)
-    } else {
-        matches
-            .remove_one::<OsString>("bwlimit")
-            .map(BandwidthArgument::Limit)
-    };
-    let excludes = matches
-        .remove_many::<OsString>("exclude")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let includes = matches
-        .remove_many::<OsString>("include")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let compare_destinations = matches
-        .remove_many::<OsString>("compare-dest")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let copy_destinations = matches
-        .remove_many::<OsString>("copy-dest")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let exclude_from = matches
-        .remove_many::<OsString>("exclude-from")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let include_from = matches
-        .remove_many::<OsString>("include-from")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let filters: Vec<OsString> = matches
-        .remove_many::<OsString>("filter")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let rsync_filter_shortcuts = rsync_filter_indices.len() as u8;
-    let filter_args = collect_filter_arguments(&filters, &filter_indices, &rsync_filter_indices);
-    let cvs_exclude = matches.get_flag("cvs-exclude");
-    let files_from = matches
-        .remove_many::<OsString>("files-from")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let from0 = matches.get_flag("from0");
-    let info = matches
-        .remove_many::<OsString>("info")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let debug = matches
-        .remove_many::<OsString>("debug")
-        .map(|values| values.collect())
-        .unwrap_or_default();
-    let ignore_existing = matches.get_flag("ignore-existing");
-    let update = matches.get_flag("update");
-    let password_file = matches.remove_one::<OsString>("password-file");
-    let protocol = matches.remove_one::<OsString>("protocol");
-    let timeout = matches.remove_one::<OsString>("timeout");
-    let contimeout = matches.remove_one::<OsString>("contimeout");
-    let out_format = matches.remove_one::<OsString>("out-format");
-    let itemize_changes = matches.get_flag("itemize-changes");
-    let no_motd = matches.get_flag("no-motd");
-
-    Ok(ParsedArgs {
-        program_name,
-        show_help,
-        show_version,
-        human_readable,
-        dry_run,
-        list_only,
-        remote_shell,
-        connect_program,
-        daemon_port,
-        remote_options,
-        rsync_path,
-        protect_args,
-        address_mode,
-        bind_address: bind_address_raw,
-        archive,
-        delete_mode,
-        delete_excluded,
-        backup,
-        backup_dir,
-        backup_suffix,
-        checksum,
-        checksum_choice,
-        checksum_choice_arg,
-        checksum_seed,
-        size_only,
-        ignore_existing,
-        ignore_missing_args,
-        update,
-        remainder,
-        bwlimit,
-        max_delete,
-        min_size,
-        max_size,
-        modify_window,
-        compress,
-        no_compress,
-        compress_level,
-        skip_compress,
-        owner,
-        group,
-        chown,
-        chmod,
-        perms,
-        super_mode,
-        times,
-        omit_dir_times,
-        omit_link_times,
-        numeric_ids,
-        hard_links,
-        sparse,
-        copy_links,
-        copy_dirlinks,
-        copy_unsafe_links: copy_unsafe_links_option,
-        keep_dirlinks,
-        safe_links,
-        devices,
-        specials,
-        relative,
-        one_file_system,
-        implied_dirs,
-        mkpath,
-        prune_empty_dirs,
-        verbosity,
-        progress: progress_setting,
-        name_level,
-        name_overridden,
-        stats,
-        partial,
-        preallocate,
-        delay_updates,
-        partial_dir,
-        temp_dir,
-        link_dests,
-        remove_source_files,
-        inplace,
-        append,
-        append_verify: append_verify_flag,
-        msgs_to_stderr,
-        itemize_changes,
-        whole_file,
-        excludes,
-        includes,
-        compare_destinations,
-        copy_destinations,
-        link_destinations,
-        exclude_from,
-        include_from,
-        filters: filter_args.clone(),
-        cvs_exclude,
-        rsync_filter_shortcuts,
-        files_from,
-        from0,
-        info,
-        debug,
-        acls,
-        xattrs,
-        no_motd,
-        password_file,
-        protocol,
-        timeout,
-        contimeout,
-        out_format,
-    })
-}
 
 /// Renders the help text describing the currently supported options.
 fn render_help(program_name: ProgramName) -> String {
@@ -2298,7 +1557,7 @@ fn parse_timeout_argument(value: &OsStr) -> Result<TransferTimeout, Message> {
     }
 }
 
-fn parse_human_readable_level(value: &OsStr) -> Result<HumanReadableMode, clap::Error> {
+pub(crate) fn parse_human_readable_level(value: &OsStr) -> Result<HumanReadableMode, clap::Error> {
     let text = value.to_string_lossy();
     let trimmed = text.trim_matches(|ch: char| ch.is_ascii_whitespace());
     let display = if trimmed.is_empty() {
@@ -2373,7 +1632,7 @@ fn parse_max_delete_argument(value: &OsStr) -> Result<u64, Message> {
     }
 }
 
-fn parse_checksum_seed_argument(value: &OsStr) -> Result<u32, Message> {
+pub(crate) fn parse_checksum_seed_argument(value: &OsStr) -> Result<u32, Message> {
     let text = value.to_string_lossy();
     let trimmed = text.trim_matches(|ch: char| ch.is_ascii_whitespace());
     let display = if trimmed.is_empty() {
@@ -3307,7 +2566,7 @@ fn parse_bandwidth_limit(argument: &OsStr) -> Result<Option<BandwidthLimit>, Mes
     }
 }
 
-fn parse_compress_level_argument(value: &OsStr) -> Result<CompressionSetting, Message> {
+pub(crate) fn parse_compress_level_argument(value: &OsStr) -> Result<CompressionSetting, Message> {
     let text = value.to_string_lossy();
     let trimmed = text.trim();
     if trimmed.is_empty() {
